requirements.txt
asgiref==3.4.1
attrs==21.2.0
autopep8==1.5.7
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
defusedxml==0.7.1
Django==3.2.7
django-allauth==0.45.0
django-environ==0.7.0
django-filter==2.4.0
django-rest-auth==0.9.5
django-utils-six==2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
drf-tracking==1.5.0
drf-yasg==1.20.0
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
jsonschema==3.2.0
MarkupSafe==2.0.1
oauthlib==3.1.1
packaging==21.0
Pillow==8.3.2
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.7
pyrsistent==0.18.0
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
swagger-spec-validator==2.7.3
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.6
